[https://nvbugs/5508536][fix] Updated _are_stop_words to correctly handle padded tokens and simplified _write_finish_reasons.
- simplified _write_finish_reasons in Torch Sampler
- fixed bug, where padded sequences in _are_stop_words were not correctly processed
- merged both test_torch_sampler.py files into one file
- updated MTPSampler to work with changes to TorchSampler
